Add unfinished branche
On this branche I wanted to add the possibility to give the teacher the possibility to change their personal info, but as it is a nice to have and the time is limited, I decided after 4 hours of trying it wasn't worth it. It would have probably take another 6 hours, because the code for getting an user out of db isn't there yet.